2-floor.py

+++ 0x00-python_variable_annotations/2-floor.py
@@ -1,7 +1,8 @@
 #!/usr/bin/env python3
 import math
 
-def floor(n: float) -> float:
+
+def floor(n: float) -> int:
     """
     type-annotated function floor which takes a float n as
     argument and returns the floor of the float.
